add integration tests for servo bus
