--- tests/integrationtests/test_bus_control.py
@@ -1,15 +1,0 @@
-from hiwonderbuslinker.bus_control import ServoBus
-import logging
-
-logger = logging.getLogger(__name__)
-
-SERVO_ID = 1
-PORT = 'COM5'
-BAUDRATE = 115200
-TIMEOUT = 1
-
-def test_get_bus_position():
-    pass
-
-def test_set_bus_position():
-    pass
